Долг линта, порция 9: последние семнадцать — долг закрыт (#858)
Наборы данных, комплекты, документы качества — девять файлов, семнадцать функций.
Храповик 17 → 0. С этой порцией закрыт весь долг: 112 → 0.

Что уехало: обращение к колонке из выражения, модель предпросмотра
материализации, проверка настройки union, поиск и группировка объектов каталога,
кандидаты базового экземпляра, разбор прочитанной картинки, проверка значения
поля, пометка распознанного, имя и поиск связки материала.

Шесть из девяти модулей уже были под юнит-тестами — тесты переехали на новые пути
и, кстати, перестали тянуть за собой React-компонент ради чистой функции.
Четырём тесты дописаны: fieldValidation (13 случаев), baseCandidates,
qualityLinkSearch, группировка в objectsByType. Всего 583 → 607.

Живьём добавлена одна проверка — на ПРОВОДКУ: `validateConstraint` уехал из
`PrimitiveInput.tsx`, сама функция под юнит-тестом, а её вызов из формы виден
только в браузере (дробное значение в поле «Цело число» обязано получить
замечание).

Две проводки живьём НЕ проверены, и это сказано в e2e/README.md: кандидаты базы —
у демо-документов подходящих кандидатов нет, пикер честно отвечает «нет», и
проверка была бы зелёной при любом коде; пометка обязательного поля проявляется
только после «Сохранить», а прогоны не сохраняют.

Остались четыре ПРЕДУПРЕЖДЕНИЯ (храповик считает ошибки): два exhaustive-deps и
два бесполезных eslint-disable у `new Function` — директивы указывают на правила,
которые в конфиге не включены. Снимать их не стал: они объясняют намерение.

Проверено: tsc -b, сборка, 607 юнит-тестов, храповик 17 → 0 и все восемь живых
прогонов (fields 16, dialogs 13, types 9, settings 8, pages 8, keyboard 11,
routing 11, shared-ui 6).
