+++ src/client/e2e/README.md
@@ -68,6 +68,9 @@
   (`naturalWidth > 0`) — отозванный или чужой объект-URL этой проверки не проходит;
 - **base-picker-reopens-clean** — то же «открылся с прежним поиском», но у пикера базового
   экземпляра: он чинится отдельно от пикера ссылки, значит и проверяется отдельно;
+- **constraint-violation-shows-message** (порция 9) — дробное значение в поле типа «Цело число»
+  получает замечание. Стережёт ПРОВОДКУ: сама `validateConstraint` уехала в свой модуль и покрыта
+  юнит-тестом, а её вызов из формы виден только в браузере;
 - **union-clearing-keeps-active-variant** — очистка значения активного варианта НЕ перебрасывает
   форму на первый вариант. Строка берётся с НЕ-первым вариантом намеренно: подмена подставляет
   именно первый, и на строке с первым дефект неотличим от исправности (на этом первая редакция
@@ -234,6 +237,25 @@ Playwright берётся из npx-кеша, браузер — из `ms-playwri
 прогонов). **Не покрыты**: `uniqueCode` (дублирование типа пишет данные — под smoke не
 берём) и `useMaxTemplateVersions` (порог версий виден только на шаблоне, где версий
 больше порога). Их перенос сверен чтением и типами.
+
+## Что порция 9 закрыла тестами, а не прогоном
+
+Последняя порция долга линта вынесла семнадцать функций из девяти файлов-компонентов (наборы
+данных, комплекты, документы качества). Шесть модулей уже были под юнит-тестами — тесты просто
+переехали на новые пути; четырём тесты дописаны (`fieldValidation`, `baseCandidates`,
+`qualityLinkSearch`, группировка в `objectsByType`).
+
+**Живьём проверена одна проводка из трёх** — вызов `validateConstraint` из формы (см. выше).
+Не проверены живьём:
+
+- кандидаты базового экземпляра (`SCOPE_TIER` / `ancestorTypeIds` / `parseBaseRef`) — у документов
+  демо-комплекта подходящих кандидатов нет, пикер честно отвечает «Нет подходящих кандидатов», и
+  проверка была бы зелёной при любом коде;
+- пометка обязательного поля (`isMissing`) — она проявляется только после нажатия «Сохранить», а
+  прогоны ничего не сохраняют.
+
+Обе функции чистые и покрыты юнит-тестами; непроверенным остаётся только то, что форма зовёт именно
+их — это стережёт компилятор типов.
 
 ## Запуск
