+++ src/client/eslint-baseline.json
@@ -1,3 +1,1 @@
-{
-  "react-refresh/only-export-components": 17
-}
+{}
