+++ src/client/src/features/document-sets/fields/index.ts
@@ -1,7 +1,9 @@
 export * from './constants';
 export * from './FileField';
 export * from './ImageField';
+export * from './checkImageResult';
 export * from './PrimitiveInput';
+export * from './fieldValidation';
 export * from './collectConstraintViolations';
 export * from './RefPickerModal';
 export * from './InstancePickerModal';
@@ -13,5 +15,6 @@ export * from './extractToCommonData';
 export * from './ExtractToCommonDataModal';
 export * from './ComplexFields';
 export * from './BaseCandidatePicker';
+export * from './baseCandidates';
 export * from './BaseInstanceChip';
 export * from './SectionRail';
