+++ src/client/src/features/quality-docs/QualityDocsPage.tsx
@@ -25,7 +25,8 @@ import type { DocumentType } from '@/shared/api/types';
 import { QualityDocForm } from './QualityDocForm';
 import { ambiguousDocNames, docFieldByTag, docNumberOf, isAmbiguous } from './docIdentity';
 import { recognizeAndUpdate } from './recognizeImported';
-import { QualityDocLinks, matchesLink } from './QualityDocLinks';
+import { QualityDocLinks } from './QualityDocLinks';
+import { matchesLink } from './qualityLinkSearch';
 import { ScopeReachNote } from './ScopeReachNote';
 import { docState, EXPIRING_SOON_DAYS, type DocState } from './docState';
 import { useRememberedSelection } from '@/shared/hooks/useRememberedSelection';
